requirements.txt
argparse==1.4.0
pytz==2022.1
paho-mqtt==1.6.1